Compiler warnings fixed. Fixed a bug in the acceleration edge for holonommic robots. The majority of index types is now changed to int in order to avoid explicit signed castings whenever index computations are performed. Missing return statements added to compound assignment operator overloads in PoseSE2 class.

+++ include/teb_local_planner/obstacles.h
@@ -500,7 +500,7 @@ public:
       // check if point is in the interior of the polygon
       // point in polygon test - raycasting (http://www.ecse.rpi.edu/Homepages/wrf/Research/Short_Notes/pnpoly.html)
       // using the following algorithm we may obtain false negatives on edge-cases, but that's ok for our purposes	
-      size_t i, j;
+      int i, j;
       bool c = false;
       for (i = 0, j = noVertices()-1; i < noVertices(); j = i++) 
       {
@@ -616,7 +616,7 @@ public:
   /**
     * @brief Get the number of vertices defining the polygon (the first vertex is counted once)
     */
-  size_t noVertices() const {return vertices_.size();}
+  int noVertices() const {return (int)vertices_.size();}
   
   
   ///@}
